Moved NotAvailableInMock to @/components/page and switched election report route to use NotAvailableInMock again
