--- frontend/src/features/election_management/components/index.ts
@@ -1,3 +1,0 @@
-export * from "./ElectionCreatePage";
-export * from "./ElectionHomePage";
-export * from "./ElectionReportPage";
