+++ frontend/src/features/election_management/routes.tsx
@@ -1,12 +1,17 @@
 import { RouteObject } from "react-router";
 
+import { NotAvailableInMock } from "@/components/page/NotAvailableInMock";
 import { t } from "@/lib/i18n";
 
-import { ElectionReportPage } from "./components";
+import { ElectionReportPage } from "./components/ElectionReportPage";
 
 export const electionManagementRoutes: RouteObject[] = [
   {
     index: true,
-    element: __API_MSW__ ? <div>{`${t("election.title.finish_data_entry")} - Abacus`}</div> : <ElectionReportPage />,
+    element: __API_MSW__ ? (
+      <NotAvailableInMock title={`${t("election.title.finish_data_entry")} - Abacus`} />
+    ) : (
+      <ElectionReportPage />
+    ),
   },
 ];
